Add files via upload
This is a straight-forward implementation of Polynomial Regression within which I have used a dataset consisting of weights and heights of different persons. The aim is to predict the height of a person (unknown/previously unseen values) given the value of its weight.
P.S : I have used the Sci-kit Learn library to use various functions effectively.